angle 파라미터 float 대응 및 카운트 환산 주석 보강
DXL_Manager::setSettingInfo 의 angle 파라미터를 uint16_t -> float 으로
변경하여 상위에서 전달한 소수점 각도가 절삭되지 않도록 함.
서브모듈 cpp_Motor_Class_v2 를 546da35 로 갱신함.

DXL_Class.cpp setSettingData_op 의 카운트 환산부에 주석 추가.
float -> int 변환에서 버림이 발생하나 최대 1카운트 오차라 허용함.

+++ DXL_Manager.h
@@ -137,7 +137,7 @@ public:
     }
     
 
-    void setSettingInfo(uint8_t gid, uint8_t sid, uint8_t dir, uint16_t angle, uint16_t initPosi, uint16_t reducer_ratio){
+    void setSettingInfo(uint8_t gid, uint8_t sid, uint8_t dir, float angle, uint16_t initPosi, uint16_t reducer_ratio){
     	if (0 < sid && sid < maxSize && dxlObjects[sid] != nullptr) { //sid = index
 			dxlObjects[sid]->setSettingInfo(gid, sid, dir, angle, initPosi, reducer_ratio);
 		}
